% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browseColors.R
\name{browseColors}
\alias{browseColors}
\title{Browse Color Names for Choosing}
\usage{
browseColors(ncol = 5, nrow = 20)
}
\arguments{
\item{ncol}{number of columns to show on each page}

\item{nrow}{number of rows to show on each page}
}
\description{
Browse Color Names for Choosing
}
\examples{
browseColors()

}

