% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacheAssign.R
\name{cacheAssign}
\alias{cacheAssign}
\alias{cacheLoad}
\title{Assign with cache mechanism.}
\usage{
cacheAssign(x, value)

cacheLoad(x)
}
\arguments{
\item{x}{variable name}

\item{value}{expression to assign to variable}
}
\value{
the same value assigned to the variable
}
\description{
Assign with cache mechanism.
}

