% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DirTree.R
\docType{data}
\name{DirTree}
\alias{DirTree}
\title{Manipulate directory tree.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
DirTree
}
\arguments{
\item{dir}{directory path to check}
}
\value{
no any return value. stop if any error.
}
\description{
Manipulate directory tree.
}
\examples{

  dirTree <- DirTree$new()                # create 'plots/'

  dirTree$enter("foo")                    # create 'plots/01_foo/'
    png(dirTree$get("rnorm10"))           # got    'plots/01_foo/01_rnorm10.png'
    plot(rnorm(10))
    dev.off()
  dirTree$leave()

  dirTree$enter("bar")                    # create 'plots/02_bar/'
    dirTree$enter("sub1")                 # create 'plots/02_bar/01_sub1/'
      jpeg(dirTree$get("sin", ".jpg"))    # got    'plots/02_bar/01_sub1/01_sin.jpg'
      plot(sin, 0, 2 * pi)
      dev.off()
    dirTree$leave()
    dirTree$enter("sub2")                 # create 'plots/02_bar/02_sub2/'
      f <- dirTree$get("mtcars", ".csv")  # got    'plots/02_bar/02_sub2/01_mtcars.csv'
      write.csv(mtcars, f)
    dirTree$leave()
  dirTree$leave()

  dirTree$exit()                          # safely exit
}
\keyword{datasets}

