% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensureDirectory.R
\name{ensureDirectory}
\alias{ensureDirectory}
\title{Ensure that directory exists.}
\usage{
ensureDirectory(dir)
}
\arguments{
\item{path}{directory path to check}
}
\value{
no any return value. stop if any error.
}
\description{
Ensure that directory exists.
}

